[리뷰 반영] 모델에서 UIKit 제거
