+++ Dayol-iOS/UI/Common/View/ToolBarContent/TextStyle/TextStyleModel.swift
@@ -5,7 +5,7 @@
 //  Created by 주성민 on 2021/04/04.
 //
 
-import UIKit
+import Foundation
 
 enum TextStyleModel {
 
@@ -61,43 +61,4 @@ enum TextStyleModel {
         }()
 
     }
-}
-
-// MARK: - NSAttribute -> TextStyleModel
-
-extension TextStyleModel {
-
-    /// NSParagraphStyle -> TextStyleModel.Alignment 전환
-    /// - default: Alignment.leading
-    static func alignment(paragraphStyle: NSParagraphStyle?) -> Alignment {
-        guard let alignment = paragraphStyle?.alignment else { return .leading }
-        switch alignment {
-        case .center: return .center
-        case .right: return .trailing
-        default: return .leading
-        }
-    }
-
-    /// NSAttirbutes -> TextStyleModel.AdditionalOptions
-    static func addtionalOptions(attributes: [NSAttributedString.Key : Any?]) -> [AdditionalOption] {
-        var additionalOption: [AdditionalOption] = []
-
-        if let font = attributes[.font] as? UIFont,
-           font.isBold {
-            additionalOption.append(.bold)
-        }
-
-        if let strikethroughStyle = attributes[.strikethroughStyle] as? Int,
-           strikethroughStyle == NSUnderlineStyle.single.rawValue {
-            additionalOption.append(.cancelLine)
-        }
-
-        if let underlineStyle = attributes[.underlineStyle] as? Int,
-           underlineStyle == NSUnderlineStyle.single.rawValue {
-            additionalOption.append(.underLine)
-        }
-
-        return additionalOption
-    }
-
 }
